small improvements + comments
